% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dep.R
\name{download_dep}
\alias{download_dep}
\title{To download all the required packages to a specified local directory}
\usage{
download_dep(
  x,
  destination_dir = "./",
  server_link = "http://cran.rstudio.com/"
)
}
\arguments{
\item{x}{A vector of the package names to be downloaded.}

\item{destination_dir}{Destination directory for the packages to be downloaded. The default is the current working directory.}

\item{server_link}{The link of the server to download the packages from. The default is the official CRAN website (http://cran.rstudio.com/).}
}
\value{
.tar.gz: All the R packages to be downloaded with the format of .tar.gz
}
\description{
This function aims to download all the required packages to a specified local directory.
}
\note{

}
\examples{
# download the dependencies, and all the dependencies of the dependencies, and so on of the package "cli".
package_vec = pkg_dep_all('cli')
# Depth: 1
# Depth: 2

package_vec
# [1] "cli"     "glue"    "utils"   "methods"
download_dep(package_vec)
}
\author{
Miao Cai <email: miao.cai@outlook.com>
}
