% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dep1.R
\name{pkg_dep1}
\alias{pkg_dep1}
\title{To query all the dependencies of an R package as a vector.}
\usage{
pkg_dep1(x)
}
\arguments{
\item{x}{A character string of the name of an R package.}
}
\value{
a vector: a unique vector of the dependencies of an R package.
}
\description{
This function query all the dependencies of an R package.
}
\note{
It only delve into one sub-level of the dependencies of an R package.
}
\examples{
# query the dependencies of the package "cli"
pkg_dep1('cli')
# [1] "glue"  "utils"
}
\author{
Miao Cai <email: miao.cai@outlook.com>
}
