% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dep_all.R
\name{pkg_dep_all}
\alias{pkg_dep_all}
\title{To recursively query all the dependencies, dependencies of the dependencies, and so on of an R package as a vector with unique elements.}
\usage{
pkg_dep_all(x)
}
\arguments{
\item{x}{A character string of the name of an R package.}
}
\value{
a vector: a unique vector of all the dependencies, dependencies of the dependencies, and so on of
}
\description{
This function recursively query all the dependencies, dependencies of the dependencies, and so on of an R package.
}
\note{
This function queries into all sub-levels of the dependencies, dependencies of the dependencies, and so on of an R package. The accompanying messages shows the progress (the depth of the query) of the query process.
}
\examples{
# query the dependencies, and all the dependencies of the dependencies, and so on of the package "cli".
pkg_dep_all('cli')
# Depth: 1
# Depth: 2
# [1] "cli"     "glue"    "utils"   "methods"
}
\references{
None.
}
\author{
Miao Cai <email: miao.cai@outlook.com>
}
