% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_vec.R
\name{pkg_dep_vec}
\alias{pkg_dep_vec}
\title{To calculate the Elixhauser Comorbidity Index}
\usage{
pkg_dep_vec(x)
}
\arguments{
\item{x}{Your data file in which Elixhauser Comorbidity Index is to be calculated}
}
\value{
a vector: a new data.frame named "data". This data frame contains a new variable "Elix_Index": The Elixhauser Comorbidity Index, developed by Anne Elixhauser in 1998
}
\description{
This file aims to calculate the Elixhauser Comorbidity Index.
}
\note{
The Elixhauser Comorbidity Index does not include age as a component of the index.
}
\author{
Miao Cai <email: miao.cai@outlook.com>
}
